rename req_categories
